replace internal sound chip clocks with double, same as cpus, and do proper rounding for sample rates, etc
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5135 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/machine/z80fmly.h
@@ -11,7 +11,7 @@
 typedef struct
 {
 	int num;                                      /* number of CTCs to emulate */
-	int baseclock[MAX_CTC];                           /* timer clock */
+	double baseclock[MAX_CTC];                    /* timer clock */
 	int notimer[MAX_CTC];                         /* timer disablers */
 	void (*intr[MAX_CTC])(int which);             /* callback when change interrupt status */
 	mem_write_handler zc0[MAX_CTC];   /* ZC/TO0 callback */
