+++ src/sound/5220intf.h
@@ -6,7 +6,7 @@
 
 struct TMS5220interface
 {
-	int baseclock;				/* clock rate = 80 * output sample rate,     */
+	double baseclock;			/* clock rate = 80 * output sample rate,     */
 								/* usually 640000 for 8000 Hz sample rate or */
 								/* usually 800000 for 10000 Hz sample rate.  */
 	int mixing_level;
